Collapse unnecessary base classes into implementations
